Split rvc to rvc and rvc_shared
