+++ src/Makefile.am
@@ -7,6 +7,6 @@ rvd_SOURCES = rvd.c log.c cmd.c util.c vpn.c
 rvd_LDADD = $(JSONC_LIBS)
 rvd_LDFLAGS = $(JSONC_LDFLAGS)
 
-rvc_SOURCES = rvc.c util.c
+rvc_SOURCES = rvc_shared.c rvc.c util.c
 rvc_LDADD = $(JSONC_LIBS)
 rvc_LDFLAGS = $(JSONC_LDFLAGS)
